requirements.txt
# RAG 系统核心依赖
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.0.12

# HTTP客户端
requests==2.31.0

# 向量数据库
chromadb==0.4.22
faiss-cpu==1.7.4

# 文本处理
sentence-transformers==2.2.2
jieba==0.42.1
pypinyin==0.49.0

# 机器学习和数据处理
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Web框架
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# 文件处理
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志和监控
loguru==0.7.2

# 开发工具
pytest==7.4.3
black==23.9.1
flake8==6.1.0
